fix(generate): update base URL for OpenAI API endpoint refactor(generate): replace go-openai with go-anthropic for LLM client integration fix(generate): correct parameter passing for new anthropic client refactor(domain): move client creation for anthropic to its own file feat(domain): add support for invoking JSON using anthropic API
